missions, glide, some other stuff

+++ app/src/main/java/com/grappim/spacexapp/model/rocket/FirstStage.kt
@@ -8,7 +8,7 @@ data class FirstStage(
   @SerializedName("engines")
   val engines: Int?,
   @SerializedName("fuel_amount_tons")
-  val fuelAmountTons: Int?,
+  val fuelAmountTons: Double?,
   @SerializedName("reusable")
   val reusable: Boolean?,
   @SerializedName("thrust_sea_level")
